+++ piccolo/apps/fixtures/commands/load.py
@@ -1,6 +1,9 @@
 from __future__ import annotations
 
+import sys
 import typing as t
+
+import typing_extensions
 
 from piccolo.apps.fixtures.commands.shared import (
     FixtureConfig,
@@ -8,12 +11,17 @@ from piccolo.apps.fixtures.commands.shared import (
 )
 from piccolo.conf.apps import Finder
 from piccolo.engine import engine_finder
+from piccolo.query.mixins import OnConflictAction
 from piccolo.table import Table, sort_table_classes
 from piccolo.utils.encoding import load_json
 from piccolo.utils.list import batch
 
 
-async def load_json_string(json_string: str, chunk_size: int = 1000):
+async def load_json_string(
+    json_string: str,
+    chunk_size: int = 1000,
+    on_conflict_action: t.Optional[OnConflictAction] = None,
+):
     """
     Parses the JSON string, and inserts the parsed data into the database.
     """
@@ -71,10 +79,23 @@ async def load_json_string(json_string: str, chunk_size: int = 1000):
             rows = data[table_class]
 
             for chunk in batch(data=rows, chunk_size=chunk_size):
-                await table_class.insert(*chunk).run()
+                query = table_class.insert(*chunk)
+                if on_conflict_action is not None:
+                    query = query.on_conflict(
+                        target=table_class._meta.primary_key,
+                        action=on_conflict_action,
+                        values=table_class._meta.columns,
+                    )
+                await query.run()
 
 
-async def load(path: str = "fixture.json", chunk_size: int = 1000):
+async def load(
+    path: str = "fixture.json",
+    chunk_size: int = 1000,
+    on_conflict: t.Optional[
+        typing_extensions.Literal["DO NOTHING", "DO UPDATE"]
+    ] = None,
+):
     """
     Reads the fixture file, and loads the contents into the database.
 
@@ -86,8 +107,28 @@ async def load(path: str = "fixture.json", chunk_size: int = 1000):
         determined by the database adapter, which has a max number of
         parameters per query.
 
+    :param on_conflict:
+        If specified, the fixture will be upserted, meaning that if a row
+        already exists with a matching primary key, then it will be overridden
+        if "DO UPDATE", or it will be ignored if "DO NOTHING".
+
     """
     with open(path, "r") as f:
         contents = f.read()
 
-    await load_json_string(contents, chunk_size=chunk_size)
+    on_conflict_action: t.Optional[OnConflictAction] = None
+
+    if on_conflict:
+        try:
+            on_conflict_action = OnConflictAction(on_conflict.upper())
+        except ValueError:
+            sys.exit(
+                f"{on_conflict} isn't a valid option - use 'DO NOTHING' or "
+                "'DO UPDATE'."
+            )
+
+    await load_json_string(
+        contents,
+        chunk_size=chunk_size,
+        on_conflict_action=on_conflict_action,
+    )