+++ piccolo/apps/migrations/commands/backwards.py
@@ -2,6 +2,7 @@ from __future__ import annotations
 
 import os
 import sys
+import typing as t
 
 from piccolo.apps.migrations.auto.migration_manager import MigrationManager
 from piccolo.apps.migrations.commands.base import (
@@ -9,6 +10,7 @@ from piccolo.apps.migrations.commands.base import (
     MigrationResult,
 )
 from piccolo.apps.migrations.tables import Migration
+from piccolo.conf.apps import AppConfig, MigrationModule
 
 
 class BackwardsMigrationManager(BaseMigrationManager):
@@ -27,20 +29,10 @@ class BackwardsMigrationManager(BaseMigrationManager):
         self.preview = preview
         super().__init__()
 
-    async def run(self) -> MigrationResult:
-        await self.create_migration_table()
-
-        app_modules = self.get_app_modules()
-
-        migration_modules = {}
-
-        for app_module in app_modules:
-            app_config = getattr(app_module, "APP_CONFIG")
-            if app_config.app_name == self.app_name:
-                migration_modules = self.get_migration_modules(
-                    app_config.migrations_folder_path
-                )
-                break
+    async def run_migrations_backwards(self, app_config: AppConfig):
+        migration_modules: t.Dict[
+            str, MigrationModule
+        ] = self.get_migration_modules(app_config.migrations_folder_path)
 
         ran_migration_ids = await Migration.get_migrations_which_ran(
             app_name=self.app_name
@@ -111,6 +103,11 @@ class BackwardsMigrationManager(BaseMigrationManager):
             message = "Not proceeding."
             print(message, file=sys.stderr)
             return MigrationResult(success=False, message=message)
+
+    async def run(self) -> MigrationResult:
+        await self.create_migration_table()
+        app_config = self.get_app_config(self.app_name)
+        return await self.run_migrations_backwards(app_config=app_config)
 
 
 async def run_backwards(