+++ tests/apps/migrations/auto/test_schema_differ.py
@@ -4,7 +4,14 @@ import typing as t
 from unittest import TestCase
 from unittest.mock import MagicMock, call, patch
 
-from piccolo.apps.migrations.auto import DiffableTable, SchemaDiffer
+from piccolo.apps.migrations.auto.schema_differ import (
+    DiffableTable,
+    RenameColumn,
+    RenameColumnCollection,
+    RenameTable,
+    RenameTableCollection,
+    SchemaDiffer,
+)
 from piccolo.columns.column_types import Numeric, Varchar
 
 
@@ -32,7 +39,7 @@ class TestSchemaDiffer(TestCase):
         self.assertTrue(len(create_tables.statements) == 1)
         self.assertEqual(
             create_tables.statements[0],
-            "manager.add_table('Band', tablename='band')",
+            "manager.add_table(class_name='Band', tablename='band', schema=None, columns=None)",  # noqa: E501
         )
 
         new_table_columns = schema_differ.new_table_columns
@@ -57,7 +64,7 @@ class TestSchemaDiffer(TestCase):
         self.assertTrue(len(schema_differ.drop_tables.statements) == 1)
         self.assertEqual(
             schema_differ.drop_tables.statements[0],
-            "manager.drop_table(class_name='Band', tablename='band')",
+            "manager.drop_table(class_name='Band', tablename='band', schema=None)",  # noqa: E501
         )
 
     def test_rename_table(self):
@@ -85,11 +92,46 @@ class TestSchemaDiffer(TestCase):
         self.assertTrue(len(schema_differ.rename_tables.statements) == 1)
         self.assertEqual(
             schema_differ.rename_tables.statements[0],
-            "manager.rename_table(old_class_name='Band', old_tablename='band', new_class_name='Act', new_tablename='act')",  # noqa
+            "manager.rename_table(old_class_name='Band', old_tablename='band', new_class_name='Act', new_tablename='act')",  # noqa: E501
         )
 
         self.assertEqual(schema_differ.create_tables.statements, [])
         self.assertEqual(schema_differ.drop_tables.statements, [])
+
+    def test_change_schema(self):
+        """
+        Testing changing the schema.
+        """
+        schema: t.List[DiffableTable] = [
+            DiffableTable(
+                class_name="Band",
+                tablename="band",
+                columns=[],
+                schema="schema_1",
+            )
+        ]
+        schema_snapshot: t.List[DiffableTable] = [
+            DiffableTable(
+                class_name="Band",
+                tablename="band",
+                columns=[],
+                schema=None,
+            )
+        ]
+
+        schema_differ = SchemaDiffer(
+            schema=schema, schema_snapshot=schema_snapshot, auto_input="y"
+        )
+
+        self.assertEqual(len(schema_differ.change_table_schemas.statements), 1)
+
+        self.assertEqual(
+            schema_differ.change_table_schemas.statements[0],
+            "manager.change_table_schema(class_name='Band', tablename='band', new_schema='schema_1', old_schema=None)",  # noqa: E501
+        )
+
+        self.assertListEqual(schema_differ.create_tables.statements, [])
+        self.assertListEqual(schema_differ.drop_tables.statements, [])
 
     def test_add_column(self):
         """
@@ -123,7 +165,7 @@ class TestSchemaDiffer(TestCase):
         self.assertTrue(len(schema_differ.add_columns.statements) == 1)
         self.assertEqual(
             schema_differ.add_columns.statements[0],
-            "manager.add_column(table_class_name='Band', tablename='band', column_name='genre', db_column_name='genre', column_class_name='Varchar', column_class=Varchar, params={'length': 255, 'default': '', 'null': False, 'primary_key': False, 'unique': False, 'index': False, 'index_method': IndexMethod.btree, 'choices': None, 'db_column_name': None, 'secret': False})",  # noqa
+            "manager.add_column(table_class_name='Band', tablename='band', column_name='genre', db_column_name='genre', column_class_name='Varchar', column_class=Varchar, params={'length': 255, 'default': '', 'null': False, 'primary_key': False, 'unique': False, 'index': False, 'index_method': IndexMethod.btree, 'choices': None, 'db_column_name': None, 'secret': False})",  # noqa: E501
         )
 
     def test_drop_column(self):
@@ -158,7 +200,7 @@ class TestSchemaDiffer(TestCase):
         self.assertTrue(len(schema_differ.drop_columns.statements) == 1)
         self.assertEqual(
             schema_differ.drop_columns.statements[0],
-            "manager.drop_column(table_class_name='Band', tablename='band', column_name='genre', db_column_name='genre')",  # noqa
+            "manager.drop_column(table_class_name='Band', tablename='band', column_name='genre', db_column_name='genre')",  # noqa: E501
         )
 
     def test_rename_column(self):
@@ -196,7 +238,7 @@ class TestSchemaDiffer(TestCase):
         self.assertEqual(
             schema_differ.rename_columns.statements,
             [
-                "manager.rename_column(table_class_name='Band', tablename='band', old_column_name='title', new_column_name='name', old_db_column_name='title', new_db_column_name='name')"  # noqa
+                "manager.rename_column(table_class_name='Band', tablename='band', old_column_name='title', new_column_name='name', old_db_column_name='title', new_db_column_name='name')"  # noqa: E501
             ],
         )
 
@@ -449,3 +491,54 @@ class TestSchemaDiffer(TestCase):
 
     def test_alter_default(self):
         pass
+
+
+class TestRenameTableCollection(TestCase):
+    collection = RenameTableCollection(
+        rename_tables=[
+            RenameTable(
+                old_class_name="Manager",
+                old_tablename="manager",
+                new_class_name="Manager1",
+                new_tablename="manager_1",
+            )
+        ]
+    )
+
+    def test_was_renamed_from(self):
+        self.assertTrue(
+            self.collection.was_renamed_from(old_class_name="Manager")
+        )
+        self.assertFalse(
+            self.collection.was_renamed_from(old_class_name="Band")
+        )
+
+    def test_renamed_from(self):
+        self.assertEqual(
+            self.collection.renamed_from(new_class_name="Manager1"), "Manager"
+        )
+        self.assertIsNone(
+            self.collection.renamed_from(new_class_name="Band"),
+        )
+
+
+class TestRenameColumnCollection(TestCase):
+    def test_for_table_class_name(self):
+        rename_column = RenameColumn(
+            table_class_name="Manager",
+            tablename="manager",
+            old_column_name="name",
+            new_column_name="full_name",
+            old_db_column_name="name",
+            new_db_column_name="full_name",
+        )
+
+        collection = RenameColumnCollection(rename_columns=[rename_column])
+
+        self.assertListEqual(
+            collection.for_table_class_name(table_class_name="Manager"),
+            [rename_column],
+        )
+        self.assertListEqual(
+            collection.for_table_class_name(table_class_name="Band"), []
+        )